+++ source/matrix.h
@@ -6,7 +6,7 @@ namespace LibRayMarching
 {
 	struct Matrix
 	{
-		double field[16];
+		float field[16];
 
 		Matrix();
 		Matrix(const Matrix& Matrix);
@@ -22,10 +22,10 @@ namespace LibRayMarching
 		Vector operator*(const Vector& vec) const;
 	};
 
-	Matrix MatrixRotate(const Vector& axis, const double& angle_radian);
+	Matrix MatrixRotate(const Vector& axis, const float& angle_radian);
 	Matrix MatrixTranslate(const Vector& translation);
 	Matrix MatrixScale(const Vector& scale);
-	Matrix MatrixScale(const double& scale);
+	Matrix MatrixScale(const float& scale);
 	Matrix MatrixInverse(const Matrix& matrix);
 
 	std::ostream& operator<< (std::ostream& stream, const Matrix& matrix);
